Introduce Unikraft-specific functionality for Approval/Reviews
This commit introduces new functionaltiy into the Concourse resource [0] to
facilitate the review process of new incoming code changes to Unikraft ecosystem
repositories.  The review process is stricter than what is provided by Github
and also requires the modification of commit messages before upstreaming or
merging and this resource as well as the changes introduced in this commit aim
at automating the process.  This is currently not possible using Github's PR
tools and so this resource acts as mechanic for triggering a PR ready for
merging based on several key properties.

The resource has two key parameters which are continiously monitored: reviewers
and approvers.  Both parties leave comments or a review entry in a PR which must
match some regular expression.  In this resource it is programmable, but in
reality for Unikraft's purpose it is strict and will always be in the format of
Git trailers:

  Reviewed-by: Your Name <your.name@email.com>
  Approved-by: Your Name <your.name@email.com>

When the minimum number of reviews and approvals are met from
commenters/reviewers who match the specified Github team, then a new version is
output to Concourse to trigger an action.  In the case of Unikraft, the action
is a merge which takes the properties (namely the matching name and email from
the trailers) and processes a merge with updated Git commit messages.

[0]: https://github.com/nderjung/concourse-github-pr-comment-resource.git

Signed-off-by: Alexander Jung <a.jung@lancs.ac.uk>